test(replay): Rename & streamline replay test modules
