+++ packages/replay/test/integration/sampling.test.ts
@@ -1,10 +1,9 @@
-// mock functions need to be imported first
 import { mockRrweb, mockSdk } from '../index';
 import { useFakeTimers } from '../utils/use-fake-timers';
 
 useFakeTimers();
 
-describe('Replay (sampling)', () => {
+describe('Integration | sampling', () => {
   it('does nothing if not sampled', async () => {
     const { record: mockRecord } = mockRrweb();
     const { replay } = await mockSdk({
@@ -19,13 +18,10 @@ describe('Replay (sampling)', () => {
 
     jest.spyOn(replay, 'loadSession');
     jest.spyOn(replay, 'addListeners');
-    // @ts-ignore private
-    expect(replay.initialState).toEqual(undefined);
     jest.runAllTimers();
 
     expect(replay.session?.sampled).toBe(false);
-    // @ts-ignore private
-    expect(replay._context).toEqual(
+    expect(replay.getContext()).toEqual(
       expect.objectContaining({
         initialTimestamp: expect.any(Number),
         initialUrl: 'http://localhost/',
