+++ packages/replay/test/unit/coreHandlers/handleFetch.test.ts
@@ -19,28 +19,30 @@ const DEFAULT_DATA = {
   startTimestamp: 10000,
 };
 
-it('formats fetch calls from core SDK to replay breadcrumbs', function () {
-  expect(handleFetch(DEFAULT_DATA)).toEqual({
-    type: 'resource.fetch',
-    name: '/api/0/organizations/sentry/',
-    start: 10,
-    end: 15,
-    data: {
-      method: 'GET',
-      statusCode: 200,
-    },
+describe('Unit | coreHandlers | handleFetch', () => {
+  it('formats fetch calls from core SDK to replay breadcrumbs', function () {
+    expect(handleFetch(DEFAULT_DATA)).toEqual({
+      type: 'resource.fetch',
+      name: '/api/0/organizations/sentry/',
+      start: 10,
+      end: 15,
+      data: {
+        method: 'GET',
+        statusCode: 200,
+      },
+    });
   });
-});
 
-it('ignores fetches that have not completed yet', function () {
-  const data = {
-    ...DEFAULT_DATA,
-  };
+  it('ignores fetches that have not completed yet', function () {
+    const data = {
+      ...DEFAULT_DATA,
+    };
 
-  // @ts-ignore: The operand of a 'delete' operator must be optional.ts(2790)
-  delete data.endTimestamp;
-  // @ts-ignore: The operand of a 'delete' operator must be optional.ts(2790)
-  delete data.response;
+    // @ts-ignore: The operand of a 'delete' operator must be optional.ts(2790)
+    delete data.endTimestamp;
+    // @ts-ignore: The operand of a 'delete' operator must be optional.ts(2790)
+    delete data.response;
 
-  expect(handleFetch(data)).toEqual(null);
+    expect(handleFetch(data)).toEqual(null);
+  });
 });
