--- packages/replay/test/unit/multiple-instances.test.ts
@@ -1,8 +1,0 @@
-import { Replay } from './../../src';
-
-it('throws on creating multiple instances', function () {
-  expect(() => {
-    new Replay();
-    new Replay();
-  }).toThrow();
-});
