+++ packages/replay/test/unit/multipleInstances.test.ts
@@ -1,0 +1,10 @@
+import { Replay } from '../../src';
+
+describe('Unit | multipleInstances', () => {
+  it('throws on creating multiple instances', function () {
+    expect(() => {
+      new Replay();
+      new Replay();
+    }).toThrow();
+  });
+});
