+++ packages/replay/test/unit/session/getSession.test.ts
@@ -8,7 +8,7 @@ import { makeSession } from '../../../src/session/Session';
 jest.mock('@sentry/utils', () => {
   return {
     ...(jest.requireActual('@sentry/utils') as { string: unknown }),
-    uuid4: jest.fn(() => 'test_session_id'),
+    uuid4: jest.fn(() => 'test_session_uuid'),
   };
 });
 
@@ -27,166 +27,168 @@ function createMockSession(when: number = new Date().getTime()) {
   });
 }
 
-beforeAll(() => {
-  jest.spyOn(CreateSession, 'createSession');
-  jest.spyOn(FetchSession, 'fetchSession');
-  WINDOW.sessionStorage.clear();
-});
-
-afterEach(() => {
-  WINDOW.sessionStorage.clear();
-  (CreateSession.createSession as jest.MockedFunction<typeof CreateSession.createSession>).mockClear();
-  (FetchSession.fetchSession as jest.MockedFunction<typeof FetchSession.fetchSession>).mockClear();
-});
-
-it('creates a non-sticky session when one does not exist', function () {
-  const { session } = getSession({
-    expiry: 900000,
-    stickySession: false,
-    ...SAMPLE_RATES,
+describe('Unit | session | getSession', () => {
+  beforeAll(() => {
+    jest.spyOn(CreateSession, 'createSession');
+    jest.spyOn(FetchSession, 'fetchSession');
+    WINDOW.sessionStorage.clear();
   });
 
-  expect(FetchSession.fetchSession).not.toHaveBeenCalled();
-  expect(CreateSession.createSession).toHaveBeenCalled();
-
-  expect(session).toEqual({
-    id: 'test_session_id',
-    segmentId: 0,
-    lastActivity: expect.any(Number),
-    sampled: 'session',
-    started: expect.any(Number),
+  afterEach(() => {
+    WINDOW.sessionStorage.clear();
+    (CreateSession.createSession as jest.MockedFunction<typeof CreateSession.createSession>).mockClear();
+    (FetchSession.fetchSession as jest.MockedFunction<typeof FetchSession.fetchSession>).mockClear();
   });
 
-  // Should not have anything in storage
-  expect(FetchSession.fetchSession()).toBe(null);
-});
+  it('creates a non-sticky session when one does not exist', function () {
+    const { session } = getSession({
+      expiry: 900000,
+      stickySession: false,
+      ...SAMPLE_RATES,
+    });
 
-it('creates a non-sticky session, regardless of session existing in sessionStorage', function () {
-  saveSession(createMockSession(new Date().getTime() - 10000));
+    expect(FetchSession.fetchSession).not.toHaveBeenCalled();
+    expect(CreateSession.createSession).toHaveBeenCalled();
 
-  const { session } = getSession({
-    expiry: 1000,
-    stickySession: false,
-    ...SAMPLE_RATES,
-  });
-
-  expect(FetchSession.fetchSession).not.toHaveBeenCalled();
-  expect(CreateSession.createSession).toHaveBeenCalled();
-
-  expect(session).toBeDefined();
-});
-
-it('creates a non-sticky session, when one is expired', function () {
-  const { session } = getSession({
-    expiry: 1000,
-    stickySession: false,
-    ...SAMPLE_RATES,
-    currentSession: makeSession({
-      id: 'old_session_id',
-      lastActivity: new Date().getTime() - 1001,
-      started: new Date().getTime() - 1001,
+    expect(session).toEqual({
+      id: 'test_session_uuid',
       segmentId: 0,
+      lastActivity: expect.any(Number),
       sampled: 'session',
-    }),
+      started: expect.any(Number),
+    });
+
+    // Should not have anything in storage
+    expect(FetchSession.fetchSession()).toBe(null);
   });
 
-  expect(FetchSession.fetchSession).not.toHaveBeenCalled();
-  expect(CreateSession.createSession).toHaveBeenCalled();
+  it('creates a non-sticky session, regardless of session existing in sessionStorage', function () {
+    saveSession(createMockSession(new Date().getTime() - 10000));
 
-  expect(session).toBeDefined();
-  expect(session.id).not.toBe('old_session_id');
-});
+    const { session } = getSession({
+      expiry: 1000,
+      stickySession: false,
+      ...SAMPLE_RATES,
+    });
 
-it('creates a sticky session when one does not exist', function () {
-  expect(FetchSession.fetchSession()).toBe(null);
+    expect(FetchSession.fetchSession).not.toHaveBeenCalled();
+    expect(CreateSession.createSession).toHaveBeenCalled();
 
-  const { session } = getSession({
-    expiry: 900000,
-    stickySession: true,
-    sessionSampleRate: 1.0,
-    errorSampleRate: 0.0,
+    expect(session).toBeDefined();
   });
 
-  expect(FetchSession.fetchSession).toHaveBeenCalled();
-  expect(CreateSession.createSession).toHaveBeenCalled();
+  it('creates a non-sticky session, when one is expired', function () {
+    const { session } = getSession({
+      expiry: 1000,
+      stickySession: false,
+      ...SAMPLE_RATES,
+      currentSession: makeSession({
+        id: 'old_session_id',
+        lastActivity: new Date().getTime() - 1001,
+        started: new Date().getTime() - 1001,
+        segmentId: 0,
+        sampled: 'session',
+      }),
+    });
 
-  expect(session).toEqual({
-    id: 'test_session_id',
-    segmentId: 0,
-    lastActivity: expect.any(Number),
-    sampled: 'session',
-    started: expect.any(Number),
+    expect(FetchSession.fetchSession).not.toHaveBeenCalled();
+    expect(CreateSession.createSession).toHaveBeenCalled();
+
+    expect(session).toBeDefined();
+    expect(session.id).not.toBe('old_session_id');
   });
 
-  // Should not have anything in storage
-  expect(FetchSession.fetchSession()).toEqual({
-    id: 'test_session_id',
-    segmentId: 0,
-    lastActivity: expect.any(Number),
-    sampled: 'session',
-    started: expect.any(Number),
-  });
-});
+  it('creates a sticky session when one does not exist', function () {
+    expect(FetchSession.fetchSession()).toBe(null);
 
-it('fetches an existing sticky session', function () {
-  const now = new Date().getTime();
-  saveSession(createMockSession(now));
+    const { session } = getSession({
+      expiry: 900000,
+      stickySession: true,
+      sessionSampleRate: 1.0,
+      errorSampleRate: 0.0,
+    });
 
-  const { session } = getSession({
-    expiry: 1000,
-    stickySession: true,
-    sessionSampleRate: 1.0,
-    errorSampleRate: 0.0,
-  });
+    expect(FetchSession.fetchSession).toHaveBeenCalled();
+    expect(CreateSession.createSession).toHaveBeenCalled();
 
-  expect(FetchSession.fetchSession).toHaveBeenCalled();
-  expect(CreateSession.createSession).not.toHaveBeenCalled();
-
-  expect(session).toEqual({
-    id: 'test_session_id',
-    segmentId: 0,
-    lastActivity: now,
-    sampled: 'session',
-    started: now,
-  });
-});
-
-it('fetches an expired sticky session', function () {
-  const now = new Date().getTime();
-  saveSession(createMockSession(new Date().getTime() - 2000));
-
-  const { session } = getSession({
-    expiry: 1000,
-    stickySession: true,
-    ...SAMPLE_RATES,
-  });
-
-  expect(FetchSession.fetchSession).toHaveBeenCalled();
-  expect(CreateSession.createSession).toHaveBeenCalled();
-
-  expect(session.id).toBe('test_session_id');
-  expect(session.lastActivity).toBeGreaterThanOrEqual(now);
-  expect(session.started).toBeGreaterThanOrEqual(now);
-  expect(session.segmentId).toBe(0);
-});
-
-it('fetches a non-expired non-sticky session', function () {
-  const { session } = getSession({
-    expiry: 1000,
-    stickySession: false,
-    ...SAMPLE_RATES,
-    currentSession: makeSession({
-      id: 'test_session_id_2',
-      lastActivity: +new Date() - 500,
-      started: +new Date() - 500,
+    expect(session).toEqual({
+      id: 'test_session_uuid',
       segmentId: 0,
+      lastActivity: expect.any(Number),
       sampled: 'session',
-    }),
+      started: expect.any(Number),
+    });
+
+    // Should not have anything in storage
+    expect(FetchSession.fetchSession()).toEqual({
+      id: 'test_session_uuid',
+      segmentId: 0,
+      lastActivity: expect.any(Number),
+      sampled: 'session',
+      started: expect.any(Number),
+    });
   });
 
-  expect(FetchSession.fetchSession).not.toHaveBeenCalled();
-  expect(CreateSession.createSession).not.toHaveBeenCalled();
+  it('fetches an existing sticky session', function () {
+    const now = new Date().getTime();
+    saveSession(createMockSession(now));
 
-  expect(session.id).toBe('test_session_id_2');
-  expect(session.segmentId).toBe(0);
+    const { session } = getSession({
+      expiry: 1000,
+      stickySession: true,
+      sessionSampleRate: 1.0,
+      errorSampleRate: 0.0,
+    });
+
+    expect(FetchSession.fetchSession).toHaveBeenCalled();
+    expect(CreateSession.createSession).not.toHaveBeenCalled();
+
+    expect(session).toEqual({
+      id: 'test_session_id',
+      segmentId: 0,
+      lastActivity: now,
+      sampled: 'session',
+      started: now,
+    });
+  });
+
+  it('fetches an expired sticky session', function () {
+    const now = new Date().getTime();
+    saveSession(createMockSession(new Date().getTime() - 2000));
+
+    const { session } = getSession({
+      expiry: 1000,
+      stickySession: true,
+      ...SAMPLE_RATES,
+    });
+
+    expect(FetchSession.fetchSession).toHaveBeenCalled();
+    expect(CreateSession.createSession).toHaveBeenCalled();
+
+    expect(session.id).toBe('test_session_uuid');
+    expect(session.lastActivity).toBeGreaterThanOrEqual(now);
+    expect(session.started).toBeGreaterThanOrEqual(now);
+    expect(session.segmentId).toBe(0);
+  });
+
+  it('fetches a non-expired non-sticky session', function () {
+    const { session } = getSession({
+      expiry: 1000,
+      stickySession: false,
+      ...SAMPLE_RATES,
+      currentSession: makeSession({
+        id: 'test_session_uuid_2',
+        lastActivity: +new Date() - 500,
+        started: +new Date() - 500,
+        segmentId: 0,
+        sampled: 'session',
+      }),
+    });
+
+    expect(FetchSession.fetchSession).not.toHaveBeenCalled();
+    expect(CreateSession.createSession).not.toHaveBeenCalled();
+
+    expect(session.id).toBe('test_session_uuid_2');
+    expect(session.segmentId).toBe(0);
+  });
 });
