+++ packages/replay/test/unit/util/isSampled.test.ts
@@ -13,10 +13,9 @@ const cases: [number, number, boolean][] = [
   [0.5, 0.0, true],
 ];
 
-jest.spyOn(Math, 'random');
-const mockRandom = Math.random as jest.MockedFunction<typeof Math.random>;
+describe('Unit | util | isSampled', () => {
+  const mockRandom = jest.spyOn(Math, 'random');
 
-describe('isSampled', () => {
   test.each(cases)(
     'given sample rate of %p and RNG returns %p, result should be %p',
     (sampleRate: number, mockRandomValue: number, expectedResult: boolean) => {
